add commands for parsing and additional end point testing
